Move api node version test to integration tests
